// gen/diff_pb.ts
export interface Symbol {
    /**
     * Name of the symbol
     *
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * Demangled name of the symbol
     *
     * @generated from protobuf field: optional string demangled_name = 2;
     */
    demangled_name?: string;
    /**
     * Symbol address
     *
     * @generated from protobuf field: uint64 address = 3;
     */
    address: bigint;
    /**
     * Symbol size
     *
     * @generated from protobuf field: uint64 size = 4;
     */
    size: bigint;
    /**
     * Bitmask of SymbolFlag
     *
     * @generated from protobuf field: uint32 flags = 5;
     */
    flags: number;
}
// ...
export interface FunctionDiff {
    /**
     * @generated from protobuf field: objdiff.diff.Symbol symbol = 1;
     */
    symbol?: Symbol;
    /**
     * @generated from protobuf field: repeated objdiff.diff.InstructionDiff instructions = 2;
     */
    instructions: InstructionDiff[];
    /**
     * @generated from protobuf field: optional float match_percent = 3;
     */
    match_percent?: number;
}
/**
 * @generated from protobuf message objdiff.diff.DataDiff
 */
export interface DataDiff {
    /**
     * @generated from protobuf field: objdiff.diff.DiffKind kind = 1;
     */
    kind: DiffKind;
    /**
     * @generated from protobuf field: bytes data = 2;
     */
    data: Uint8Array;
    /**
     * May be larger than data
     *
     * @generated from protobuf field: uint64 size = 3;
     */
    size: bigint;
}
/**
 * @generated from protobuf message objdiff.diff.SectionDiff
 */
export interface SectionDiff {
    /**
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * @generated from protobuf field: objdiff.diff.SectionKind kind = 2;
     */
    kind: SectionKind;
    /**
     * @generated from protobuf field: uint64 size = 3;
     */
    size: bigint;
    /**
     * @generated from protobuf field: uint64 address = 4;
     */
    address: bigint;
    /**
     * @generated from protobuf field: repeated objdiff.diff.FunctionDiff functions = 5;
     */
    functions: FunctionDiff[];
    /**
     * @generated from protobuf field: repeated objdiff.diff.DataDiff data = 6;
     */
    data: DataDiff[];
    /**
     * @generated from protobuf field: optional float match_percent = 7;
     */
    match_percent?: number;
}
/**
 * @generated from protobuf message objdiff.diff.ObjectDiff
 */
export interface ObjectDiff {
    /**
     * @generated from protobuf field: repeated objdiff.diff.SectionDiff sections = 1;
     */
    sections: SectionDiff[];
}
// ...
export enum SectionKind {
    /**
     * @generated from protobuf enum value: SECTION_UNKNOWN = 0;
     */
    SECTION_UNKNOWN = 0,
    /**
     * @generated from protobuf enum value: SECTION_TEXT = 1;
     */
    SECTION_TEXT = 1,
    /**
     * @generated from protobuf enum value: SECTION_DATA = 2;
     */
    SECTION_DATA = 2,
    /**
     * @generated from protobuf enum value: SECTION_BSS = 3;
     */
    SECTION_BSS = 3,
    /**
     * @generated from protobuf enum value: SECTION_COMMON = 4;
     */
    SECTION_COMMON = 4
}
// @generated message type with reflection information, may provide speed optimized methods
class Symbol$Type extends MessageType<Symbol> {
    constructor() {
        super("objdiff.diff.Symbol", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "demangled_name", kind: "scalar", localName: "demangled_name", opt: true, T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "address", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 4, name: "size", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 5, name: "flags", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<Symbol>): Symbol {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.address = 0n;
        message.size = 0n;
        message.flags = 0;
        if (value !== undefined)
            reflectionMergePartial<Symbol>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Symbol): Symbol {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* optional string demangled_name */ 2:
                    message.demangled_name = reader.string();
                    break;
                case /* uint64 address */ 3:
                    message.address = reader.uint64().toBigInt();
                    break;
                case /* uint64 size */ 4:
                    message.size = reader.uint64().toBigInt();
                    break;
                case /* uint32 flags */ 5:
                    message.flags = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Symbol, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* optional string demangled_name = 2; */
        if (message.demangled_name !== undefined)
            writer.tag(2, WireType.LengthDelimited).string(message.demangled_name);
        /* uint64 address = 3; */
        if (message.address !== 0n)
            writer.tag(3, WireType.Varint).uint64(message.address);
        /* uint64 size = 4; */
        if (message.size !== 0n)
            writer.tag(4, WireType.Varint).uint64(message.size);
        /* uint32 flags = 5; */
        if (message.flags !== 0)
            writer.tag(5, WireType.Varint).uint32(message.flags);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message objdiff.diff.Symbol
 */
export const Symbol = new Symbol$Type();
// ...
export const FunctionDiff = new FunctionDiff$Type();
// ...
export const DataDiff = new DataDiff$Type();
// ...
export const SectionDiff = new SectionDiff$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ObjectDiff$Type extends MessageType<ObjectDiff> {
    constructor() {
        super("objdiff.diff.ObjectDiff", [
            { no: 1, name: "sections", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => SectionDiff }
        ]);
    }
    create(value?: PartialMessage<ObjectDiff>): ObjectDiff {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sections = [];
        if (value !== undefined)
            reflectionMergePartial<ObjectDiff>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ObjectDiff): ObjectDiff {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated objdiff.diff.SectionDiff sections */ 1:
                    message.sections.push(SectionDiff.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ObjectDiff, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated objdiff.diff.SectionDiff sections = 1; */
        for (let i = 0; i < message.sections.length; i++)
            SectionDiff.internalBinaryWrite(message.sections[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message objdiff.diff.ObjectDiff
 */
export const ObjectDiff = new ObjectDiff$Type();